requirements.txt
fastapi
uvicorn[standard]
grpcio
grpcio-tools
aioredis
httpx
etcd3
pydantic
kopf
kubernetes
prometheus-client
opentelemetry-api
opentelemetry-sdk
python-json-logger==2.0.3
